Move heartbeat control to separate file
